@@ -12,29 +12,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.9",
-    "@angular/cdk": "^15.2.9",
-    "@angular/common": "^15.2.9",
-    "@angular/compiler": "^15.2.9",
-    "@angular/core": "^15.2.9",
-    "@angular/forms": "^15.2.9",
-    "@angular/material": "^15.2.9",
-    "@angular/platform-browser": "^15.2.9",
-    "@angular/platform-browser-dynamic": "^15.2.9",
-    "@angular/router": "^15.2.9",
+    "@angular/animations": "^16.0.0",
+    "@angular/cdk": "^16.0.0",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/material": "^16.0.0",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
     "@ngneat/until-destroy": "^9.2.3",
-    "@ngrx/effects": "^15.4.0",
-    "@ngrx/entity": "^15.4.0",
-    "@ngrx/store": "^15.4.0",
-    "@ngrx/store-devtools": "^15.4.0",
+    "@ngrx/effects": "^16.0.0-rc.0",
+    "@ngrx/entity": "^16.0.0-rc.0",
+    "@ngrx/store": "^16.0.0-rc.0",
+    "@ngrx/store-devtools": "^16.0.0-rc.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.5.0",
     "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.8",
-    "@angular/cli": "~15.2.8",
-    "@angular/compiler-cli": "^15.2.9",
+    "@angular-devkit/build-angular": "^16.0.0",
+    "@angular/cli": "~16.0.0",
+    "@angular/compiler-cli": "^16.0.0",
     "@types/jasmine": "~4.3.1",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.2",
